:truck: Move engine resources to own submodule
